--- src/os_v5/drivers/environment.ts
@@ -1,10 +1,0 @@
-
-export enum EnvironmentName {
-  mmo = "mmo",
-  sim = "sim",
-  swc = "swc",
-  botarena = "botarena",
-  private = "private",
-  mockSeason = "mockSeason", /// シーズンマッチ未開催時
-  unknown = "unknown",
-}
