+++ src/os_v5/memory.ts
@@ -3,7 +3,6 @@ import { AnySerializable } from "./utility/types"
 export type KernelMemory = {
   version: string
   readonly systemCall: { [Key: string]: AnySerializable }
-  readonly driver: { [Key: string]: AnySerializable }
 }
 
 declare global {
